+++ module-core/src/plumcp/core/support/http_client_java.clj
@@ -68,6 +68,8 @@
              http-version
              http-proxy
              http-redirect
+             ssl-context    ; must be javax.net.ssl.SSLContext or nil
+             ssl-params     ; must be javax.net.ssl.SSLParameters or nil
              traffic-logger
              timeout-millis]
       :or {authenticator (Authenticator/getDefault)
@@ -95,6 +97,10 @@
                              cb] (.version cb http-version))
         f-http-proxy    (fn [^HttpClient$Builder
                              cb] (.proxy cb http-proxy))
+        f-ssl-context   (fn [^HttpClient$Builder
+                             cb] (.sslContext cb ssl-context))
+        f-ssl-params    (fn [^HttpClient$Builder
+                             cb] (.sslParameters cb ssl-params))
         f-conn-timeout  (fn [^HttpClient$Builder
                              cb] (->> (Duration/ofMillis timeout-millis)
                                       (.connectTimeout cb)))
@@ -107,6 +113,8 @@
       timeout-millis (f-conn-timeout)
       http-redirect  (f-http-redirect)
       http-proxy     (f-http-proxy)
+      ssl-context    (f-ssl-context)
+      ssl-params     (f-ssl-params)
       :and-finally   (f-builder))))